requirements.txt
bs4
requests
pillow
xmltodict
psycopg2-binary
mcstatus